+++ includes/AnyCommentCore.php
@@ -20,7 +20,7 @@ class AnyCommentCore {
 	 *
 	 * @var string
 	 */
-	public $version = '0.0.76';
+	public $version = '0.0.77';
 
 	/**
 	 * @var Pool