+++ includes/AnyCommentLoader.php
@@ -32,6 +32,7 @@ class AnyCommentLoader {
 
 		// Other
 		'AnyComment\Admin\AnyCommentWPComments',
+		'AnyComment\AnyCommentAvatars',
 
 		// Hooks
 		'AnyComment\Hooks\AnyCommentCommonHooks',