+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 * Fixed issue when Mail.Ru did not show as active when enabled
 * Fixed issue when some websites had problem with loading comments or widget in sidebar, #327
 * Fixed issue when user was not notified about new reply when sent from admin panel, #330
-* Possible fix of invalid nonce error when dealing with cache plugins (implemented for: WP Super Cache, WP Rocket, WP Fastest Cache)
+* Possible fix of invalid nonce error when dealing with cache plugins (implemented for: WP Super Cache, WP Rocket, WP Fastest Cache and W3 Total Cache)
 
 ## 0.0.75 – 05.12.2018